Implemented listening on the changes in database.
* Emu implementation.
* Net implementation.
* Fix nasty bug in defer.bridge_result
* Change the way couchdb is intialized, add the
  local.ini.template not to nest very long config inside
  the source.
* Broken execution chain during agent start up and fixed the testcase.
* Agent listens on changes on his descriptor and suacides when they happen.
* Fix bug in disconnection procedure for messaging.
* Expose methods to alter descriptor in simulation driver.
* Partial pool from monitoring branch with implementation of
  AgencyAgent.terminate_hard() and bugfixes around it.

+++ src/feat/process/couchdb.py
@@ -1,5 +1,6 @@
 import shutil
 import os
+import copy
 
 from feat.process import base
 from feat.agents.base import replay
@@ -12,7 +13,7 @@ class Process(base.Base):
         self.config = dict()
         workspace = self.get_tmp_dir()
         self.config['workspace'] = workspace
-        self.config['dir'] = os.path.join(workspace, 'couch_db')
+        self.config['tempdir'] = os.path.join(workspace, 'couch_db')
         self.config['port'] = self.get_free_port()
         self.config['log'] = os.path.join(workspace, 'couch_test.log')
         self.config['local_ini'] = os.path.join(workspace, 'local.ini')
@@ -20,30 +21,31 @@ class Process(base.Base):
 
     @replay.side_effect
     def prepare_workspace(self):
-        local_ini_tmpl = format_block("""
-        [couchdb]
-        database_dir = %(dir)s
-        view_index_dir = %(dir)s
+        shutil.rmtree(self.config['tempdir'], ignore_errors=True)
 
-        [httpd]
-        port = %(port)d
-
-        [log]
-        file = %(log)s
-        """)
-        shutil.rmtree(self.config['dir'], ignore_errors=True)
+        tmpl_path = os.path.join(os.path.dirname(__file__),
+                                 'local.ini.template')
+        handle = open(tmpl_path)
 
         f = file(self.config['local_ini'], 'w')
-        f.write(local_ini_tmpl % self.config)
+        f.write(handle.read() % self.config)
         f.close()
+        os.mkdir(self.config['tempdir'])
+        os.mkdir(os.path.join(self.config['tempdir'], 'lib'))
+        os.mkdir(os.path.join(self.config['tempdir'], 'log'))
 
     def initiate(self):
         self.configure()
         self.prepare_workspace()
 
         self.command = '/usr/bin/couchdb'
-        self.env['HOME'] = os.environ['HOME']
+        self.env = copy.deepcopy(os.environ)
         self.args = ['-a', self.config['local_ini']]
+        self.keep_workdir = False
+
+    def terminate(self, keep_workdir=False):
+        self.keep_workdir = keep_workdir
+        return base.Base.terminate(self)
 
     @replay.side_effect
     def started_test(self):
@@ -53,5 +55,6 @@ class Process(base.Base):
 
     @replay.side_effect
     def on_finished(self, e):
-        shutil.rmtree(self.get_config()['workspace'], ignore_errors=True)
+        if not self.keep_workdir:
+            shutil.rmtree(self.get_config()['workspace'], ignore_errors=True)
         base.Base.on_finished(self, e)
